Improving jxst_CFStringPrepareTokenRangesArray().

+++ JXLDStringTokenUtilities.h
@@ -8,4 +8,4 @@
 
 #import <Foundation/Foundation.h>
 
-int jxst_CFStringPrepareTokenRangesArray(CFStringRef string, CFRange tokenizerRange, CFOptionFlags tokenizerOptions, CFRange **ranges);
+size_t jxst_CFStringPrepareTokenRangesArray(CFStringRef string, CFRange tokenizerRange, CFOptionFlags tokenizerOptions, CFRange **ranges);
